requirements.txt
jinja2>=3.1.0
weasyprint>=62.0
python-pptx>=0.6.23
